Test: api endpoint 변경으로 컨트롤러 테스트코드 추가
